factor out open discussions preview
