+++ app/components/building-blocks/open-discussion-preview.tsx
@@ -1,0 +1,41 @@
+import { Link } from '@remix-run/react'
+import moment from 'moment'
+import { type FrontPagePost } from '#app/modules/posts/post-types.ts'
+import { PostContent } from './post-content.tsx'
+
+export function OpenDiscussionPreview({
+	post,
+	className,
+}: {
+	post: FrontPagePost
+	className?: string
+}) {
+	const ageString = moment(post.createdAt).fromNow()
+	const commentString = post.nTransitiveComments == 1 ? 'comment' : 'comments'
+
+	return (
+		<div
+			className={
+				'mb-2 w-full min-w-0 rounded-xl border-2 border-solid border-gray-200 bg-post px-3 py-2 dark:border-gray-700 ' +
+				(className || '')
+			}
+		>
+			<div className="flex">
+				<div className="flex w-full flex-col">
+					<div className="mb-2 text-sm opacity-50">{ageString}</div>
+					<PostContent
+						content={post.content}
+						maxLines={2}
+						deactivateLinks={false}
+						linkTo={`/post/${post.id}`}
+					/>
+					<div className="mt-auto text-sm opacity-50">
+						<Link to={`/post/${post.id}`}>
+							{post.nTransitiveComments} {commentString}
+						</Link>
+					</div>
+				</div>
+			</div>
+		</div>
+	)
+}
